+++ frontend/src/metabase/query_builder/components/view/ConvertQueryModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ConvertQueryModal";
